requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
python-dotenv==1.0.1
email-validator==2.1.0.post1
Werkzeug==3.0.1
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
redis>=4.5.1

# Aggiunte per sicurezza
python-jose[cryptography]==3.3.0
argon2-cffi==21.3.0
pyOpenSSL==23.2.0
pytest==7.4.3
pytest-cov==4.1.0
